Make the quantum register amplitude human readable
Move from a 1D array of length 2^N to a 2x2x2… dimensional array. Factorising out the state is now trivial.

For a register Q which contains three qubits, calculating the probability that the first qubit is a zero is equal to | Q[0,:,:] |^2. This is much more human readable and allows partial measurement of the register simpler to implement.

Once gate application functionality is added again, this should be profiled to check the performance difference (will require much more reshaping of the amplitudes)